--- a/cmds/test.js
+++ b/cmds/test.js
@@ -1,8 +1,5 @@
 'use strict'
 
-const test = require('../src/test')
-const onError = require('../src/error-handler')
-
 module.exports = {
   command: 'test',
   desc: 'Test your code in different environments',
@@ -53,6 +50,8 @@ module.exports = {
     }
   },
   handler (argv) {
+    const test = require('../src/test')
+    const onError = require('../src/error-handler')
     test.run(argv).catch(onError)
   }
 }