This makes the CLI a lot faster as aegir only have to load required commands
when it's actually needed, not always. Every command gets faster as well
since it only loads what needed (node tests don't need to load Karma
anymore for example)

This branch:
  `time node cli.js` => ~0.2 seconds
  `time node cli.js test` => ~1.6 seconds

Current master:
  `time node cli.js` => ~1.0 seconds
  `time node cli.js test` => ~2.1 seconds